Revert back to a toggle for the mouse layer
Automatically toggling the mouse layer is a bit buggy (e.g. sometimes it
disables the layer when it shouldn't), and the use of timers makes the
whole experience a bit of a hit and miss. Having to press a toggle
button is a tad annoying, but at least it's less annoying.